# server/swagger-doc/swagger-api.json
{"openapi":"3.0.0","paths":{"/auth/register":{"post":{"operationId":"register","summary":"Register a new user","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":"User successfully registered."},"400":{"description":"Invalid request data. Please provide valid user information."},"409":{"description":"A user with this username or email already exists"},"500":{"description":"Internal server error occurred. Please try again later."}},"tags":["Authentication"]}},"/auth/signIn":{"post":{"operationId":"signIn","summary":"Sign in to the system","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignInDto"}}}},"responses":{"200":{"description":"User signed in successfully."},"401":{"description":"Invalid credentials. Please check your email and password."},"409":{"description":"A user with this username or email already exists"},"500":{"description":"Internal server error occurred. Please try again later."}},"tags":["Authentication"]}},"/auth/forgotPassword":{"post":{"operationId":"forgotPassword","summary":"Initiate password reset","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ForgotPasswordDto"}}}},"responses":{"200":{"description":"Reset code verified successfully."},"404":{"description":"User not found."},"500":{"description":"Failed to send password reset code. Please try again later.d."}},"tags":["Authentication"]}},"/auth/verifyResetCode":{"post":{"operationId":"verifyResetCode","summary":"Verify password reset code","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VerifyResetCodeDto"}}}},"responses":{"200":{"description":"Reset code verified successfully."},"404":{"description":"Not Found - User not found."},"406":{"description":"Not Acceptable - Verification code incorrect."},"500":{"description":"Internal Server Error - Failed to verify reset code. Please try again later."}},"tags":["Authentication"]}},"/auth/resetPassword":{"post":{"operationId":"resetPassword","summary":"Reset password","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResetPasswordDto"}}}},"responses":{"200":{"description":"Password reset successfully."},"404":{"description":"Not Found - User not found."},"500":{"description":"Internal Server Error - Failed to reset password. Please try again later."}},"tags":["Authentication"]}},"/users/me":{"get":{"operationId":"getMe","summary":"Fetch current user (own) profile","parameters":[],"responses":{"200":{"description":"Return the current user profile."},"401":{"description":"Unauthorized"},"404":{"description":"User profile cannot be found"},"500":{"description":"Internal Server Error - Failed to get user profile. Please try again later."}},"tags":["Users"]}},"/users/me/update":{"patch":{"operationId":"update","summary":"Update current user (own) profile","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":"Profile update successfully"},"400":{"description":"User provided invalid username and/or email"},"500":{"description":"The server is experiencing an error. "}},"tags":["Users"]}},"/users/me/delete":{"delete":{"operationId":"delete","summary":"Delete current user (own) account","parameters":[],"responses":{"200":{"description":"Profile deleted successfully"},"404":{"description":"User not found"},"500":{"description":"The server is experiencing an error. "}},"tags":["Users"]}},"/locations":{"post":{"operationId":"create","summary":"Create a new location","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateLocationDto"}}}},"responses":{"201":{"description":"Location created successfully."},"409":{"description":"Conflict - Location already exists."},"500":{"description":"Internal Server Error - Failed to create location."}},"tags":["Locations"]},"get":{"operationId":"findAll","summary":"Get all locations","parameters":[],"responses":{"200":{"description":"Locations found successfully."},"500":{"description":"Internal Server Error - Failed to retrieve locations. Please try again later."}},"tags":["Locations"]}},"/locations/id/{id}":{"get":{"operationId":"findOneById","summary":"Get a location by ID","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Location found successfully."},"404":{"description":"Not Found - Location with the specified ID cannot be found."},"500":{"description":"Internal Server Error - Failed to retrieve locations. Please try again later."}},"tags":["Locations"]}},"/locations/name/{name}":{"get":{"operationId":"findOneByName","summary":"Get a location by name","parameters":[{"name":"name","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Location found successfully."},"404":{"description":"Not Found - Location with the specified name cannot be found."},"500":{"description":"Internal Server Error - Failed to retrieve locations. Please try again later."}},"tags":["Locations"]}},"/locations/code/{code}":{"get":{"operationId":"findOneByShortCode","summary":"Get a location by short code","parameters":[{"name":"code","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Location found successfully."},"404":{"description":"Not Found - Location with the specified short code cannot be found."},"500":{"description":"Internal Server Error - Failed to retrieve locations. Please try again later."}},"tags":["Locations"]}},"/locations/{id}":{"patch":{"operationId":"update","summary":"Update a location","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateLocationDto"}}}},"responses":{"200":{"description":"Location updated successfully."},"404":{"description":"Not Found - Location with the specified ID cannot be found."},"500":{"description":"Internal Server Error - Failed to update location. Please try again later."}},"tags":["Locations"]},"delete":{"operationId":"remove","summary":"Delete a location","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Location successfully deleted."},"404":{"description":"Not Found - Location with the specified ID cannot be found."},"500":{"description":"Internal Server Error - Failed to delete location. Please try again later."}},"tags":["Locations"]}}},"info":{"title":"VolunteerLink API endpoints","description":"Documentation for RESTful API endpoints of VolunteerLink project","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"CreateUserDto":{"type":"object","properties":{"firstName":{"type":"string","description":"First name of the user"},"lastName":{"type":"string","description":"Last name of the user"},"username":{"type":"string","description":"Username of the user"},"email":{"type":"string","description":"Email address of the user"},"password":{"type":"string","description":"User password","minLength":8,"maxLength":20},"locationId":{"type":"string","description":"ID of the user's location"}},"required":["firstName","lastName","username","email","password","locationId"]},"SignInDto":{"type":"object","properties":{"email":{"type":"string","description":"User email"},"password":{"type":"string","description":"User password"}},"required":["email","password"]},"ForgotPasswordDto":{"type":"object","properties":{"email":{"type":"string","description":"Email address of the user for password reset"}},"required":["email"]},"VerifyResetCodeDto":{"type":"object","properties":{"resetCode":{"type":"string","description":"Reset code sent to the user by email for password reset"},"email":{"type":"string","description":"Email address of the user for password reset"}},"required":["resetCode","email"]},"ResetPasswordDto":{"type":"object","properties":{"email":{"type":"string","description":"Email address of the user"},"newPassword":{"type":"string","description":"New password not less than 8 character","minLength":8,"maxLength":20}},"required":["email","newPassword"]},"UpdateUserDto":{"type":"object","properties":{"firstName":{"type":"string","description":"First name of the user"},"lastName":{"type":"string","description":"Last name of the user"},"username":{"type":"string","description":"Username of the user"},"email":{"type":"string","description":"Email address of the user"},"bio":{"type":"string","description":"Bio text of the user"},"locationId":{"type":"string","description":"ID of the user's location"},"locationPreference":{"type":"string","description":"Location preference of the user","enum":["IN_PERSON","REMOTE","BOTH"]},"timePreference":{"type":"string","description":"Time preference of the user","enum":["SHORT_TERM","LONG_TERM","BOTH"]}}},"CreateLocationDto":{"type":"object","properties":{"name":{"type":"string","description":"Name of the location"},"code":{"type":"string","description":"Code of the location"}},"required":["name","code"]},"UpdateLocationDto":{"type":"object","properties":{"name":{"type":"string","description":"Name of the location"},"code":{"type":"string","description":"Code of the location"}}}}},"externalDocs":{"description":"Postman Collection","url":"/api-json"}}